% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datashield.assign.R
\name{datashield.assign.table}
\alias{datashield.assign.table}
\title{Table assignment}
\usage{
datashield.assign.table(conns, symbol, table, variables = NULL,
  missings = FALSE, identifiers = NULL)
}
\arguments{
\item{conns}{\code{\link{DSConnection-class}} object or a list of \code{\link{DSConnection-class}}s.}

\item{symbol}{Name of the R symbol.}

\item{table}{Fully qualified name of a table in the data repository (must be the same in each Opal).}

\item{variables}{List of variable names or Javascript expression that selects the variables of a table (ignored if value does not refere to a table). See javascript documentation: http://wiki.obiba.org/display/OPALDOC/Variable+Methods}

\item{missings}{If TRUE, missing values will be pushed from Opal to R, default is FALSE. Ignored if value is an R expression.}

\item{identifiers}{Name of the identifiers mapping to use when assigning entities to R (from Opal 2.0).}
}
\description{
Assign a table to a R symbol in the Datashield R session.
}
\examples{
\dontrun{
# assign a list of variables from table HOP of opal object o
datashield.assign.table(o, symbol="D", table="demo.HOP", variables=list("GENDER","LAB_GLUC"))

# assign all the variables matching 'LAB' from table HOP of opal object o
datashield.assign.table(o, symbol="D", table="demo.HOP", variables="name().matches('LAB_')")
}
}
