% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datashield.build.login.data.frame.o.R
\name{datashield.build.login.data.frame.o}
\alias{datashield.build.login.data.frame.o}
\title{Builds a dataframe to login to datashield}
\usage{
datashield.build.login.data.frame.o(data.computers.name,
  data.computers.url, data.computers.table.name, users.id, users.password)
}
\arguments{
\item{data.computers.name}{A vector of characters listing all the names of the data computers}

\item{data.computers.url}{A vector of characters listing each data computer HTTP address. The format is https://[TCPIP address or host name][:port]}

\item{data.computers.table.name}{A vector of characters listing the name of the table stored in a data computer}

\item{users.id}{A vector of characters listing a valid user name to log on on each server.}

\item{users.password}{A vector of characters listing the password for each user to log in to a data computer.}
}
\value{
a data frame formatter in this manner: (server,url,user,password,table). If the arguments are not correct. Then a data.frame with no rows is created.

The expactactions are as follow:
Expectation no 0: the return value is a data.frame
Expectation no 1: the number of columns is equal 5.
Expectation no 2: the number of rows is equal to the number of servers
Expectation no 3: the number of rows is equal to 0, if the length of url, user, or table is smaller than the length of server
Expectation no 4: the number of row is 0, if any of the urls does not start with http
}
\description{
This function generates a valid data frame, that can be used to login
to some data computers (i.e. opal servers). The data frame models a double-entry table. The
columns are defined as server, url, user, password and table name. Each row holds the information
in relation to one data computer. The values for each column are passed  to the function as arguments.
}
\author{
Patricia Ryser-Welch
}
