% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datashield.errorMessages.R
\name{datashield.errorMessages}
\alias{datashield.errorMessages}
\title{datashield.errorMessages}
\usage{
datashield.errorMessages()
}
\value{
NULL if no errors are found, otherwise prints the errors.
}
\description{
Retrieve and display DataSHIELD errors in CLI format
}
\details{
This function retrieves the last errors occurred in a DataSHIELD session
and displays them in a formatted manner using bullet points.
}
