% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datashield.errors.R
\name{datashield.errors}
\alias{datashield.errors}
\title{List R last errors}
\usage{
datashield.errors()
}
\description{
Get the R last errors available after the datashield.assign or datashield.aggregate calls in the Datashield R session.
}
