% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datashield.logout.R
\name{datashield.logout}
\alias{datashield.logout}
\title{Logout from DataSHIELD R sessions}
\usage{
datashield.logout(conns, save = NULL)
}
\arguments{
\item{conns}{\code{\link{DSConnection-class}} object or a list of \code{\link{DSConnection-class}}s.}

\item{save}{Save datashield sessions on each DataSHIELD data repository (if feature is supported) with provided ID (must be a character string).}
}
\description{
Clear the Datashield R sessions and logout from DataSHIELD data repositories.
}
