% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datashield.status.R
\name{datashield.method_status}
\alias{datashield.method_status}
\title{Status of the DataSHIELD methods}
\usage{
datashield.method_status(conns, type = "aggregate")
}
\arguments{
\item{conns}{\code{\link{DSConnection-class}} object or a list of \code{\link{DSConnection-class}}s.}

\item{type}{Type of the method: "aggregate" (default) or "assign".}
}
\value{
Methods availability on each server.
}
\description{
Get the status of the DataSHIELD methods in the different data repositories to check if any method is missing.
}
