% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datashield.status.R
\name{datashield.methods}
\alias{datashield.methods}
\title{List of DataSHIELD methods}
\usage{
datashield.methods(conns, type = "aggregate")
}
\arguments{
\item{conns}{\code{\link{DSConnection-class}} object or a list of \code{\link{DSConnection-class}}s.}

\item{type}{Type of the method: "aggregate" (default) or "assign".}
}
\value{
Methods details from all the servers.
}
\description{
Get the list of all the DataSHIELD methods from the different data repositories.
}
