% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datashield.status.R
\name{datashield.pkg_check}
\alias{datashield.pkg_check}
\title{Check server-side package minimum version}
\usage{
datashield.pkg_check(
  conns,
  name,
  version,
  env = getOption("datashield.env", globalenv())
)
}
\arguments{
\item{conns}{\code{\link{DSConnection-class}} object or a list of \code{\link{DSConnection-class}}s.}

\item{name}{The name of the server-side package.}

\item{version}{The minimum package version number to be matched.}

\item{env}{Environment where the package status result should be cached. Try to get it from the
'datashield.env' option, with default to the Global Environment.}
}
\description{
Check for each of the server, accessible through provided \code{\link{DSConnection-class}} objects, whether the installed
}
