% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datashield.status.R
\name{datashield.pkg_status}
\alias{datashield.pkg_status}
\title{Status of the DataSHIELD packages}
\usage{
datashield.pkg_status(conns)
}
\arguments{
\item{conns}{A list of \code{\link{DSConnection-class}}s.}
}
\value{
Packages status for each server.
}
\description{
Get the status of the DataSHIELD packages in the different data repositories to check if any package is missing.
}
