% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datashield.status.R
\name{datashield.profiles}
\alias{datashield.profiles}
\title{List of DataSHIELD profiles}
\usage{
datashield.profiles(conns)
}
\arguments{
\item{conns}{\code{\link{DSConnection-class}} object or a list of \code{\link{DSConnection-class}}s.}
}
\value{
Profiles details from all the servers.
}
\description{
Get the list of all the DataSHIELD profiles from the different data repositories: available ones
and currently applied to each connection.
}
