% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datashield.status.R
\name{datashield.resource_status}
\alias{datashield.resource_status}
\title{Status of some resources}
\usage{
datashield.resource_status(conns, resource)
}
\arguments{
\item{conns}{\code{\link{DSConnection-class}} object or a list of \code{\link{DSConnection-class}}s.}

\item{resource}{Fully qualified name of a resource in the data repository (can be a vector or must be
the same in each data repository); or a named list of fully qualified resource names (one per server
name); or a data frame with 'server' and 'resource' columns (such as the one that is used in
\code{\link{datashield.login}})}
}
\value{
Resource status for each server.
}
\description{
Get whether some identified resources are accessible in each of the data repositories.
}
