% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datashield.list.R
\name{datashield.resources}
\alias{datashield.resources}
\title{List of the resources}
\usage{
datashield.resources(conns)
}
\arguments{
\item{conns}{\code{\link{DSConnection-class}} object or a list of \code{\link{DSConnection-class}}s.}
}
\value{
Resource unique names from all the servers.
}
\description{
Get the list of all the resources from the different data repositories.
}
\examples{
\dontrun{
  datashield.resources(conns)
}
}
