% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datashield.symbol.R
\name{datashield.rm}
\alias{datashield.rm}
\title{Remove a R symbol}
\usage{
datashield.rm(conns, symbol)
}
\arguments{
\item{conns}{\code{\link{DSConnection-class}} object or a list of \code{\link{DSConnection-class}}s.}

\item{symbol}{Name of the R symbol.}
}
\description{
Remove a symbol from the current Datashield session.
}
