% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datashield.symbol.R
\name{datashield.symbols}
\alias{datashield.symbols}
\title{List R symbols}
\usage{
datashield.symbols(conns)
}
\arguments{
\item{conns}{\code{\link{DSConnection-class}} object or a list of \code{\link{DSConnection-class}}s.}
}
\description{
Get the R symbols available after the datashield.assign calls in the Datashield R session.
}
