% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datashield.status.R
\name{datashield.table_status}
\alias{datashield.table_status}
\title{Status of some tables}
\usage{
datashield.table_status(conns, tables)
}
\arguments{
\item{conns}{A list of \code{\link{DSConnection-class}}s.}

\item{tables}{A named list of fully qualified table names, per servername.}
}
\value{
Table status for each server.
}
\description{
Get whether some identified tables are accessible in each of the data repositories.
}
