% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datashield.list.R
\name{datashield.tables}
\alias{datashield.tables}
\title{List of the tables}
\usage{
datashield.tables(conns)
}
\arguments{
\item{conns}{\code{\link{DSConnection-class}} object or a list of \code{\link{DSConnection-class}}s.}
}
\value{
Table unique names from all the servers.
}
\description{
Get the list of all the tables from the different data repositories.
}
\examples{
\dontrun{
  datashield.tables(conns)
}
}
