% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datashield.workspace.R
\name{datashield.workspace_restore}
\alias{datashield.workspace_restore}
\title{Restore saved workspace to the current DataSHIELD R session}
\usage{
datashield.workspace_restore(conns, ws)
}
\arguments{
\item{conns}{\code{\link{DSConnection-class}} object or a list of \code{\link{DSConnection-class}}s.}

\item{ws}{The workspace name}
}
\description{
Restore the state of a previously saved DataSHIELD R session (the workspace saved with \code{\link{datashield.workspace_save}}) 
with the provided name from each data repository. Note that when restoring a workspace, any existing
symbol or file with same name will be overridden.
}
