% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datashield.workspace.R
\name{datashield.workspace_rm}
\alias{datashield.workspace_rm}
\title{Remove a DataSHIELD workspace}
\usage{
datashield.workspace_rm(conns, ws)
}
\arguments{
\item{conns}{\code{\link{DSConnection-class}} object or a list of \code{\link{DSConnection-class}}s.}

\item{ws}{The workspace name}
}
\description{
Remove in each data repository the workspace with the provided name.
}
