% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datashield.workspace.R
\name{datashield.workspaces}
\alias{datashield.workspaces}
\title{List saved DataSHIELD R workspaces}
\usage{
datashield.workspaces(conns)
}
\arguments{
\item{conns}{\code{\link{DSConnection-class}} object or a list of \code{\link{DSConnection-class}}s.}
}
\description{
Get the list of R workspaces that were saved during a Datashield R session.
}
