% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.appendError}
\alias{.appendError}
\title{Append error message to last errors vector}
\usage{
.appendError(name, msg, env = getOption("datashield.env", globalenv()))
}
\description{
Append error message to last errors vector
}
\keyword{internal}
