% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.asNamedListOfConnections}
\alias{.asNamedListOfConnections}
\title{Makes a single \code{\link{DSConnection-class}} object or a named list of \code{\link{DSConnection-class}} objects.}
\usage{
.asNamedListOfConnections(conns)
}
\description{
Makes a single \code{\link{DSConnection-class}} object or a named list of \code{\link{DSConnection-class}} objects.
}
\keyword{internal}
