% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.asNamedListOfResources}
\alias{.asNamedListOfResources}
\title{Makes a typical logindata data frame a list of resources named by the server in which they are defined.
Makes a character vector of resource names a list named by the connections.}
\usage{
.asNamedListOfResources(conns, value)
}
\description{
Makes a typical logindata data frame a list of resources named by the server in which they are defined.
Makes a character vector of resource names a list named by the connections.
}
\keyword{internal}
