% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.asNamedListOfTables}
\alias{.asNamedListOfTables}
\title{Makes a typical logindata data frame a list of tables named by the server in which they are defined.
Makes a character vector of table names a list named by the connections.}
\usage{
.asNamedListOfTables(conns, value)
}
\description{
Makes a typical logindata data frame a list of tables named by the server in which they are defined.
Makes a character vector of table names a list named by the connections.
}
\keyword{internal}
