% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.asNamedListOfValues}
\alias{.asNamedListOfValues}
\title{Makes a typical logindata data frame a list of items named by the server in which they are defined.
Makes a character vector of item names a list named by the connections.}
\usage{
.asNamedListOfValues(conns, value, colname)
}
\description{
Makes a typical logindata data frame a list of items named by the server in which they are defined.
Makes a character vector of item names a list named by the connections.
}
\keyword{internal}
