% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.checkLastErrors}
\alias{.checkLastErrors}
\title{Check if there are last errors}
\usage{
.checkLastErrors(env = getOption("datashield.env", globalenv()))
}
\description{
Check if there are last errors
}
\keyword{internal}
