% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datashield.connections.R
\name{.getDSConnections}
\alias{.getDSConnections}
\title{Gets the \code{\link{DSConnection-class}} objects}
\usage{
.getDSConnections(env)
}
\arguments{
\item{env}{The environment where to search for the connection symbols.}
}
\value{
A list of \code{\link{DSConnection-class}} objects obtained after login into the servers
}
\description{
The function searches for a list containing object of type \code{\link{DSConnection-class}}
in the current environment; if more than one list is found it return the lastest.
This way no matter what the user calls his connection object list it will be captured.
}
\keyword{internal}
