% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.getSleepTime}
\alias{.getSleepTime}
\title{Get time to sleep depending on the numer of previous iterations}
\usage{
.getSleepTime(checks)
}
\description{
Get time to sleep depending on the numer of previous iterations
}
\keyword{internal}
