% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.hasLastErrors}
\alias{.hasLastErrors}
\title{Check if there are last errors}
\usage{
.hasLastErrors(name, env = getOption("datashield.env", globalenv()))
}
\description{
Check if there are last errors
}
\keyword{internal}
