% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findDSConnections.R
\name{.isDSConnection}
\alias{.isDSConnection}
\title{Check if provided object is a S4 class instance and if this class inherits from \code{\link{DSConnection-class}}.}
\usage{
.isDSConnection(obj)
}
\description{
Check if provided object is a S4 class instance and if this class inherits from \code{\link{DSConnection-class}}.
}
\keyword{internal}
