% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datashield.errors.R
\name{.remove_curly_brackets}
\alias{.remove_curly_brackets}
\title{Remove Curly Brackets from Strings}
\usage{
.remove_curly_brackets(errors)
}
\arguments{
\item{errors}{A list of strings that may contain curly brackets.}
}
\value{
A list of strings with curly brackets replaced by parentheses.
}
\description{
This function replaces all curly brackets in the input strings with parentheses.
}
