% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSConnection.R
\name{dsAggregate}
\alias{dsAggregate}
\title{Aggregate data}
\usage{
dsAggregate(conn, expr)
}
\arguments{
\item{conn}{An object that inherits from \code{\link{DSConnection-class}}.}

\item{expr}{Expression to evaluate.}
}
\description{
Aggregate some data from the DataSHIELD R session using a valid R expression. The aggregation expression
must satisfy the data repository's DataSHIELD configuration.

\Sexpr[results=rd,stage=render]{DSI:::methods_as_rd("dsAggregate")}
}
\examples{
\dontrun{
con <- dbConnect(DSOpal::Opal(), "username", "password", "https://opal.example.org")
dsAssignTable(con, "D", "test.CNSIM")
dsAggregate(con, as.symbol("meanDS(D$WEIGHT)"))
dsDisconnect(con)
}
}
\seealso{
Other DSConnection generics: \code{\link{DSConnection-class}},
  \code{\link{dsAssignExpr}}, \code{\link{dsAssignTable}},
  \code{\link{dsDisconnect}}, \code{\link{dsGetInfo}},
  \code{\link{dsHasTable}}, \code{\link{dsIsAsync}},
  \code{\link{dsListSymbols}}, \code{\link{dsListTables}},
  \code{\link{dsRmSymbol}}
}
\concept{DSConnection generics}
