% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSConnection.R
\name{dsAssignTable}
\alias{dsAssignTable}
\title{Assign a remote table}
\usage{
dsAssignTable(conn, symbol, table, variables = NULL, missings = FALSE,
  identifiers = NULL)
}
\arguments{
\item{conn}{An object that inherits from \code{\link{DSConnection-class}}.}

\item{symbol}{Name of the R symbol.}

\item{table}{Fully qualified name of a table in the data repository.}

\item{variables}{List of variable names or Javascript expression that selects the variables of a table (ignored if value does not refere to a table). See javascript documentation: http://wiki.obiba.org/display/OPALDOC/Variable+Methods}

\item{missings}{If TRUE, missing values will be pushed from Opal to R, default is FALSE. Ignored if value is an R expression.}

\item{identifiers}{Name of the identifiers mapping to use when assigning entities to R.}
}
\description{
Assign a remote table from the data repository to a symbol in the DataSHIELD R session.
The table to be assigned must exist for the DataSHIELD user.

\Sexpr[results=rd,stage=render]{DSI:::methods_as_rd("dsAssignTable")}
}
\examples{
\dontrun{
con <- dbConnect(DSOpal::Opal(), "server1", "username", "password", "https://opal.example.org")
dsAssignTable(con, "D", "test.CNSIM")
dsDisconnect(con)
}
}
\seealso{
Other DSConnection generics: \code{\link{DSConnection-class}},
  \code{\link{dsAggregate}}, \code{\link{dsAssignExpr}},
  \code{\link{dsDisconnect}}, \code{\link{dsGetInfo}},
  \code{\link{dsHasTable}}, \code{\link{dsIsAsync}},
  \code{\link{dsListMethods}},
  \code{\link{dsListPackages}},
  \code{\link{dsListSymbols}}, \code{\link{dsListTables}},
  \code{\link{dsListWorkspaces}}, \code{\link{dsRmSymbol}},
  \code{\link{dsRmWorkspace}},
  \code{\link{dsSaveWorkspace}}
}
\concept{DSConnection generics}
