% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSDriver.R
\docType{class}
\name{DSDriver-class}
\alias{DSDriver-class}
\title{DSDriver class}
\description{
Base class for all DataSHIELD-aware data repositories drivers (e.g., Opal, ...).
The virtual class `DSDriver` defines the operations for creating
connections.
}
\seealso{
Other DS classes: 
\code{\link{DSConnection-class}},
\code{\link{DSObject-class}},
\code{\link{DSResult-class}},
\code{\link{DSSession-class}}

Other DSDriver generics: 
\code{\link{dsConnect}()},
\code{\link{dsGetInfo}()}
}
\concept{DS classes}
\concept{DSDriver generics}
