% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSConnection.R
\name{dsHasResource}
\alias{dsHasResource}
\title{Check remote resource exists}
\usage{
dsHasResource(conn, resource)
}
\arguments{
\item{conn}{An object that inherits from \code{\link{DSConnection-class}}.}

\item{resource}{the resource fully qualified name}
}
\description{
Check if a remote resource reference exists in the data repository. Returns a logical indicating the existence of a
remote resource accessible through this connection.

\Sexpr[results=rd,stage=render]{DSI:::methods_as_rd("dsHasResource")}
}
\examples{
\dontrun{
con <- dbConnect(DSOpal::Opal(), "server1",
  "username", "password", "https://opal.example.org")
dsHasResource(con, "test.CNSIM")
dsDisconnect(con)
}
}
\seealso{
Other DSConnection generics: \code{\link{DSConnection-class}},
  \code{\link{dsAggregate}}, \code{\link{dsAssignExpr}},
  \code{\link{dsAssignResource}},
  \code{\link{dsAssignTable}}, \code{\link{dsDisconnect}},
  \code{\link{dsGetInfo}}, \code{\link{dsHasTable}},
  \code{\link{dsIsAsync}}, \code{\link{dsListMethods}},
  \code{\link{dsListPackages}},
  \code{\link{dsListResources}},
  \code{\link{dsListSymbols}}, \code{\link{dsListTables}},
  \code{\link{dsListWorkspaces}}, \code{\link{dsRmSymbol}},
  \code{\link{dsRmWorkspace}},
  \code{\link{dsSaveWorkspace}}
}
\concept{DSConnection generics}
