% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSI-package.R
\docType{package}
\name{DSI-package}
\alias{DSI}
\alias{DSI-package}
\title{DSI: 'DataSHIELD' Interface}
\description{
'DataSHIELD' is an infrastructure and series of R packages that 
    enables the remote and 'non-disclosive' analysis of sensitive research data. 
    This package defines the API that is to be implemented by 'DataSHIELD' compliant 
    data repositories.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://datashield.ac.uk}
  \item Report bugs at \url{https://github.com/datashield/DSI}
}

}
\author{
\strong{Maintainer}: Yannick Marcon \email{yannick.marcon@obiba.org} (\href{https://orcid.org/0000-0003-0138-2023}{ORCID})

Other contributors:
\itemize{
  \item Amadou Gaye (\href{https://orcid.org/0000-0002-1180-2792}{ORCID}) [contributor]
  \item Paul Burton [contributor]
}

}
\keyword{internal}
