% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSConnection.R
\name{dsListTables}
\alias{dsListTables}
\title{List remote tables}
\usage{
dsListTables(conn)
}
\arguments{
\item{conn}{an object that inherits from \code{\link{DSConnection-class}}.}
}
\description{
List remote tables from the data repository. Returns the unquoted names of remote tables
accessible through this connection.

\Sexpr[results=rd,stage=render]{DSI:::methods_as_rd("dsListTables")}
}
\examples{
\dontrun{
con <- dbConnect(DSOpal::Opal(), "username", "password", "https://opal.example.org")
dsListTables(con)
dsDisconnect(con)
}
}
\seealso{
Other DSConnection generics: \code{\link{DSConnection-class}},
  \code{\link{dsAggregate}}, \code{\link{dsAssignExpr}},
  \code{\link{dsAssignTable}}, \code{\link{dsDisconnect}},
  \code{\link{dsGetInfo}}, \code{\link{dsHasTable}},
  \code{\link{dsIsAsync}}, \code{\link{dsListSymbols}},
  \code{\link{dsRmSymbol}}
}
\concept{DSConnection generics}
