% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSLoginBuilder.R
\docType{class}
\name{DSLoginBuilder}
\alias{DSLoginBuilder}
\title{DataSHIELD login details builder}
\format{A R6 object of class DSLoginBuilder}
\usage{
DSLoginBuilder
}
\description{
Helper class for creating a valid data frame that can be used to perform \code{\link{datashield.login}}.
See also \link{newDSLoginBuilder}.
}
\section{Fields}{

\describe{
\item{\code{logins}}{A valid login details data frame to initiate the builder, optional.}

\item{\code{.silent}}{Do not warn user when non secure HTTP urls are encountered. Default is FALSE.}
}}

\section{Methods}{

\code{$new(logins, .silent)} Create new DSLoginBuilder instance with the arguments described in the Fields section.

\code{$append(server, url, table, driver, user, password, token, options)} Append a row in the login data frame with the specifications:
\code{server} is the server name, \code{url} is the url to connect to the server or a R symbol name, \code{table} is the table path
in that identifies the dataset in the server, \code{driver} is the \code{\link{DSDriver-class}} name to build the \code{\link{DSConnection-class}}, \code{user}
and \code{password} are the user credentials, \code{token} is the personal access token (ignored when user credentials are not empty),
\code{options} any options (R code to be parsed) that could be relevant for the DS connection object.

\code{$build()} Get the created login details data frame.
}

\keyword{datasets}
