% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSResult.R
\docType{class}
\name{DSResult-class}
\alias{DSResult-class}
\title{DSResult class}
\description{
This virtual class describes the result and state of execution of
a DataSHIELD request (aggregation or assignment).
}
\section{Implementation notes}{

Individual drivers are free to allow single or multiple
active results per connection.

The default show method displays a summary of the query using other
DS generics.
}

\seealso{
Other DS classes: 
\code{\link{DSConnection-class}},
\code{\link{DSDriver-class}},
\code{\link{DSObject-class}}

Other DSResult generics: 
\code{\link{dsFetch}()},
\code{\link{dsGetInfo}()},
\code{\link{dsIsCompleted}()}
}
\concept{DS classes}
\concept{DSResult generics}
