% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSSession.R
\docType{class}
\name{DSSession-class}
\alias{DSSession-class}
\title{DSSession class}
\description{
This virtual class describes the state of the R session.
}
\details{
The default show method displays a summary of the R session state using other
DS generics.
}
\seealso{
Other DS classes: 
\code{\link{DSConnection-class}},
\code{\link{DSDriver-class}},
\code{\link{DSObject-class}},
\code{\link{DSResult-class}}

Other DSSession generics: 
\code{\link{dsIsReady}()},
\code{\link{dsStateMessage}()}
}
\concept{DS classes}
\concept{DSSession generics}
