% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datashield.assign.R
\name{datashield.assign}
\alias{datashield.assign}
\title{Data assignment}
\usage{
datashield.assign(conns, symbol, value, variables = NULL,
  missings = FALSE, identifiers = NULL, async = TRUE)
}
\arguments{
\item{conns}{\code{\link{DSConnection-class}} object or a list of \code{\link{DSConnection-class}}s.}

\item{symbol}{Name of the R symbol.}

\item{value}{Fully qualified name of a variable or a table reference in data repositories
(must be the same in each data repository) or a R expression with allowed assign functions calls.}

\item{variables}{List of variable names or Javascript expression that selects the variables of a table
(ignored if value does not refere to a table). See javascript documentation:
\url{http://opaldoc.obiba.org/en/latest/magma-user-guide/variable/}}

\item{missings}{If TRUE, missing values will be pushed from data repository to R, default is FALSE.
Ignored if value is an R expression.}

\item{identifiers}{Name of the identifiers mapping to use when assigning entities to R (if supported
by data repository).}

\item{async}{Whether the result of the call should be retrieved asynchronously (TRUE means that calls are parallelized over
the connections, when the connection supports that feature, with an extra overhead of requests).}
}
\description{
Assign a table or an expression result to a R symbol in the Datashield R session.
}
\examples{
\dontrun{
# assign a list of variables from table HOP
datashield.assign(conn, symbol="D", value="demo.HOP",
  variables=list("GENDER","LAB_GLUC"))

# assign all the variables matching 'LAB' from table HOP
datashield.assign(conn, symbol="D", value="demo.HOP",
  variables="name().matches('LAB_')")
}
}
