% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datashield.build.login.data.frame.o.R
\name{datashield.build.login.data.frame.o}
\alias{datashield.build.login.data.frame.o}
\title{Builds a dataframe to login to datashield}
\usage{
datashield.build.login.data.frame.o(
  servers.name,
  servers.url,
  servers.table,
  users.name = NULL,
  users.password = NULL,
  users.token = NULL,
  servers.driver = NULL,
  .silent = FALSE
)
}
\arguments{
\item{servers.name}{A vector of characters listing all the names of the data servers.}

\item{servers.url}{A vector of characters listing each data server address:  can be a HTTP address (format is
http[s]://[TCPIP address or host name][:port]) or a R symbol name.}

\item{servers.table}{A vector of characters listing the name of the table stored in a data server.}

\item{users.name}{A vector of characters listing a valid user name to log on each server.}

\item{users.password}{A vector of characters listing the password for each user to log in to a data server.}

\item{users.token}{A vector of characters listing the API access tokens for each user to log in to a data server (ignored if user name and password are specified).}

\item{servers.driver}{A vector of characters listing the name of the DS connection driver to be used, see \code{\link{DSDriver-class}}.
When NULL this vector fallbacks to a vector of "OpalDriver" strings.}

\item{.silent}{Do not warn user when non secure HTTP urls are encountered. Default is FALSE.}
}
\value{
A data frame formatter in this manner: (server, url, driver, user, password, token, table). If the arguments are not correct. Then a data.frame with no rows is created.
}
\description{
This function generates a valid data frame, that can be used to login to some data servers implementing
the \code{\link{DSConnection-class}}. Each row of the returned data frame holds the information
in relation to one data computer: location of the server and of the dataset, connection class, credentials.
See also \link{newDSLoginBuilder}.
}
