% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datashield.errors.R
\name{datashield.errors}
\alias{datashield.errors}
\title{datashield.errors}
\usage{
datashield.errors()
}
\value{
NULL if no errors are found, otherwise prints the errors.
}
\description{
Retrieve and display the last errors occurred in a DataSHIELD session.
}
\details{
This function retrieves the last errors occurred in a DataSHIELD session
and displays them in a formatted manner using bullet points.
}
