% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datashield.status.R
\name{datashield.table_status}
\alias{datashield.table_status}
\title{Status of some tables}
\usage{
datashield.table_status(conns, table)
}
\arguments{
\item{conns}{\code{\link{DSConnection-class}} object or a list of \code{\link{DSConnection-class}}s.}

\item{table}{Fully qualified name of a table in the data repository (can be a vector or must be
the same in each data repository); or a named list of fully qualified table names (one per server
name); or a data frame with 'server' and 'table' columns (such as the one that is used in
\code{\link{datashield.login}})}
}
\value{
Table status for each server.
}
\description{
Get whether some identified tables are accessible in each of the data repositories.
}
