% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findDSConnections.R
\name{findDSConnections}
\alias{findDSConnections}
\title{Searches for \code{\link{DSConnection-class}} objects in the environment}
\usage{
findDSConnections()
}
\value{
returns a list of \code{\link{DSConnection-class}} objects or stops the process
}
\description{
If the user does not set the argument 'datasources', this function
is called to searches for opal login objects in the environment. If more than one
login object is found a prompt asks the user to choose one and if none is found
the process stops.
}
